+++ src/org/rascalmpl/library/vis/interaction/Choice.java
@@ -74,7 +74,7 @@ public class Choice extends Figure {
 	}
 
 	@Override
-	public void bbox(float desiredWidth, float desiredHeight) {
+	public void bbox(double desiredWidth, double desiredHeight) {
 		width = choice.getWidth();
 		height = choice.getHeight();
 	}
@@ -90,7 +90,7 @@ public class Choice extends Figure {
 	}
 
 	@Override
-	public void draw(float left, float top) {
+	public void draw(double left, double top) {
 		this.setLeft(left);
 		this.setTop(top);
 		fpa.setBackground(new Color(getFillColorProperty()));
