+++ src/org/rascalmpl/library/vis/interaction/TextField.java
@@ -98,7 +98,7 @@ public class TextField extends Figure {
 	}
 
 	@Override
-	public void bbox(float desiredWidth, float desiredHeight) {
+	public void bbox(double desiredWidth, double desiredHeight) {
 		width = getWidthProperty();
 		height = getHeightProperty();
 		
@@ -139,7 +139,7 @@ public class TextField extends Figure {
 	}
 
 	@Override
-	public void draw(float left, float top) {
+	public void draw(double left, double top) {
 		this.setLeft(left);
 		this.setTop(top);
 		fpa.setBackground(new Color(getFillColorProperty()));
